@@ -1,15 +1,15 @@
 {
-  "name": "ui-router-react",
-  "version": "0.4.0",
+  "name": "@uirouter/react",
+  "version": "0.5.0",
   "description": "State based routing for React",
   "main": "lib/index.js",
   "typings": "lib/index.d.ts",
   "jsnext:main": "lib-esm/index.js",
   "scripts": {
     "test": "jest --watch",
-    "test:ci": "jest",
-    "dev": "NODE_ENV=development webpack-dev-server --host 0.0.0.0 --port 8000 --config webpack.example.js --history-api-fallback --content-base example",
-    "build": "npm run clean && tsc && tsc -m es6 --outDir lib-esm && NODE_ENV=production webpack",
+    "test:ci": "NODE_ENV=production jest",
+    "dev": "cross-env NODE_ENV=development webpack-dev-server --host 0.0.0.0 --port 8000 --config webpack.example.js --history-api-fallback --content-base example",
+    "build": "npm run clean && tsc && tsc -m es6 --outDir lib-esm && cross-env NODE_ENV=production webpack",
     "clean": "shx rm -rf _bundles lib lib-esm build",
     "docs": "./scripts/docs.js",
     "package": "npm run build"
@@ -32,23 +32,30 @@
     "url": "https://github.com/ui-router/react.git"
   },
   "dependencies": {
-    "classnames": "^2.2.5",
-    "es6-shim": "^0.35.1",
-    "ui-router-core": "=4.0.0"
+    "@uirouter/core": "=5.0.3",
+    "classnames": "^2.2.5"
+  },
+  "peerDependencies": {
+    "prop-types": "^15.5.8",
+    "react": "^15.0.0 || ^16.0.0-alpha"
   },
   "devDependencies": {
     "@types/classnames": "0.0.31",
+    "@types/lodash": "^4.14.64",
+    "@types/prop-types": "^15.5.1",
     "@types/react": "^0.14.43",
     "@types/react-dom": "^0.14.18",
-    "awesome-typescript-loader": "3.0.0-beta.17",
+    "awesome-typescript-loader": "^3.1.3",
     "babel-jest": "^15.0.0",
     "babel-polyfill": "^6.13.0",
     "babel-preset-es2015": "^6.14.0",
     "babel-preset-react": "^6.11.1",
     "conventional-changelog-cli": "1.1.1",
-    "enzyme": "^2.4.1",
+    "cross-env": "^4.0.0",
+    "enzyme": "^2.8.2",
     "glob": "^7.0.5",
-    "jest": "^17.0.3",
+    "jest": "^20.0.3",
+    "prop-types": "^15.5.10",
     "react": "^15.0.2",
     "react-addons-test-utils": "^15.3.1",
     "react-dom": "^15.0.2",
@@ -56,16 +63,16 @@
     "replace-in-file": "^2.0.3",
     "shelljs": "^0.7.0",
     "shx": "^0.1.4",
-    "sinon": "^1.17.5",
-    "ts-jest": "^17.0.0",
+    "sinon": "^2.3.1",
+    "ts-jest": "^20.0.4",
     "typedoc": "0.5.3",
     "typedoc-plugin-external-module-name": "^1.0.8",
     "typedoc-plugin-internal-external": "^1.0.8",
     "typedoc-plugin-ui-router": "^1.0.6",
-    "typescript": "^2.1.0",
+    "typescript": "^2.3.3",
     "ui-router-typedoc-themes": "^1.0.2",
-    "webpack": "^1.13.0",
-    "webpack-dev-server": "^1.14.1"
+    "webpack": "^2.3.1",
+    "webpack-dev-server": "^2.4.2"
   },
   "jest": {
     "rootDir": "src",
@@ -79,6 +86,13 @@
       "json"
     ],
     "testURL": "http://localhost:4000",
-    "testRegex": "/__tests__/.*\\.(ts|tsx|js)$"
+    "testRegex": "/__tests__/.*\\.(ts|tsx|js)$",
+    "globals": {
+      "__TS_CONFIG__": {
+        "jsx": "react",
+        "module": "commonjs",
+        "target": "es6"
+      }
+    }
   }
 }
