@@ -1,17 +1,13 @@
 /**
- * # Components
- * 
- * React Components and their APIs
- *
- * @preferred
  * @reactapi
  * @module components
  */ /** */
 import * as React from 'react';
-import {Component, PropTypes, Children} from 'react';
+import {Component, Children} from 'react';
+import * as PropTypes from 'prop-types';
 import {UIRouterReact, ReactStateDeclaration} from '../index';
-import {UIRouterPlugin} from 'ui-router-core';
-import {servicesPlugin} from 'ui-router-core';
+import {UIRouterPlugin} from '@uirouter/core';
+import {servicesPlugin} from '@uirouter/core';
 
 export interface UIRouterProps {
   plugins?: any[]; // should fix type
