@@ -2,11 +2,11 @@
  * @reactapi
  * @module react
  */ /** */
-import {Component} from "react";
-import {StateDeclaration, _ViewDeclaration} from "ui-router-core";
-import {ParamDeclaration} from "ui-router-core";
-import {IInjectable} from "ui-router-core";
-import {Transition} from "ui-router-core";
+import {Component, ReactElement, StatelessComponent, ComponentClass, ClassicComponentClass} from "react";
+import {StateDeclaration, _ViewDeclaration} from "@uirouter/core";
+import {ParamDeclaration} from "@uirouter/core";
+import {IInjectable} from "@uirouter/core";
+import {Transition} from "@uirouter/core";
 
 
 /**
@@ -177,7 +177,7 @@ export interface ReactViewDeclaration extends _ViewDeclaration {
    * }
    * ```
    */
-  component?: any;
+  component?: StatelessComponent<any> | ComponentClass<any> | ClassicComponentClass<any>;
 
   /**
    * @hidden
